+++ src/main/java/nl/nn/testtool/echo2/reports/ReportsComponent.java
@@ -803,7 +803,10 @@ public class ReportsComponent extends BaseComponent implements BeanParent, Actio
 	}
 
 	public void displayReports(boolean metadataNamesChanged) {
-		if (testTool.messageCapturerWaitingForClose()) {
+		if (testTool.warnReportsInProgress()) {
+			displayError("One or more reports are in progress for more than 5 minuts");
+		}
+		if (testTool.warnMessageCapturerWaitingForClose()) {
 			displayError("One or more reports are finished but waiting for more than 30 seconds for one or more message capturers to close");
 		}
 		Storage storage = getSelectedView().getStorage();